Feature/add due date task and complete (#34)
* added due date setting of the task item

* added task can be completed
