+++ src/main/java/co/jht/exception/TaskNotFoundException.java
@@ -1,0 +1,7 @@
+package co.jht.exception;
+
+public class TaskNotFoundException extends RuntimeException {
+    public TaskNotFoundException(String message) {
+        super(message);
+    }
+}
