+++ src/main/java/co/jht/repository/TaskRepository.java
@@ -1,6 +1,6 @@
 package co.jht.repository;
 
-import co.jht.entity.TaskItem;
+import co.jht.entity.tasks.TaskItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
