+++ src/main/java/co/jht/serializer/ZonedDateTimeDeserializer.java
@@ -1,0 +1,20 @@
+package co.jht.serializer;
+
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JsonDeserializer;
+
+import java.io.IOException;
+import java.time.ZonedDateTime;
+
+import static co.jht.util.DateTimeFormatterUtil.getFormatter;
+
+public class ZonedDateTimeDeserializer extends JsonDeserializer<ZonedDateTime> {
+
+    @Override
+    public ZonedDateTime deserialize(JsonParser p, DeserializationContext ctxt)
+            throws IOException, JsonProcessingException {
+        return ZonedDateTime.parse(p.getText(), getFormatter());
+    }
+}
