+++ src/main/java/co/jht/service/TaskService.java
@@ -1,7 +1,8 @@
 package co.jht.service;
 
-import co.jht.entity.TaskItem;
+import co.jht.entity.tasks.TaskItem;
 
+import java.time.ZonedDateTime;
 import java.util.List;
 
 public interface TaskService {
@@ -9,4 +10,6 @@ public interface TaskService {
     TaskItem updateTask(TaskItem task);
     void deleteTask(Long taskId);
     List<TaskItem> getTasksByUserId(Long userId);
+    TaskItem updateTaskDueDate(Long taskId, ZonedDateTime dueDate);
+    TaskItem updateTaskCompleteStatus(Long taskId, boolean completeStatus);
 }
