+++ src/main/java/co/jht/service/impl/TaskServiceImpl.java
@@ -1,6 +1,7 @@
 package co.jht.service.impl;
 
-import co.jht.entity.TaskItem;
+import co.jht.entity.tasks.TaskItem;
+import co.jht.exception.TaskNotFoundException;
 import co.jht.repository.TaskRepository;
 import co.jht.service.TaskService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,6 +10,7 @@ import org.springframework.stereotype.Service;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.util.List;
+import java.util.Optional;
 
 import static co.jht.constants.ApplicationConstants.ASIA_TOKYO;
 
@@ -49,5 +51,32 @@ public class TaskServiceImpl implements TaskService {
                     }
                 })
                 .toList();
+    }
+
+    @Override
+    public TaskItem updateTaskDueDate(Long taskId, ZonedDateTime dueDate) {
+        Optional<TaskItem> taskOptional = taskRepository.findById(taskId);
+
+        if (taskOptional.isPresent()) {
+            TaskItem task = taskOptional.get();
+            task.setDueDate(dueDate);
+
+            return taskRepository.save(task);
+        } else {
+            throw new TaskNotFoundException("Task not found with id: " + taskId);
+        }
+    }
+
+    public TaskItem updateTaskCompleteStatus(Long taskId, boolean completeStatus) {
+        Optional<TaskItem> taskOptional = taskRepository.findById(taskId);
+
+        if (taskOptional.isPresent()) {
+            TaskItem task = taskOptional.get();
+            task.setComplete(completeStatus);
+
+            return taskRepository.save(task);
+        } else {
+            throw new TaskNotFoundException("Task not found with id: " + taskId);
+        }
     }
 }
